The MoMo transfer code
Most of us have already suffered at least one attack from scammers.
This leads to discontent subscribers and faced with that sad reality that some scammers use the system of MoMo transfer to achieve their ends, it was important to find a solution that would prevent them. 
The MoMo transfer code is a tool to provide more security to subscribers and to the system MoMo.